+++ src/sgpp/parallel/datadriven/basis/modlinear/operation/OperationMultipleEvalIterativeHybridX86SimdOCLModLinear.cpp
@@ -126,8 +126,8 @@ namespace sg {
 #ifdef _OPENMP
         } else {
 #endif
-          int worksize = static_cast<int>((storageSize - gpu_partition) / (num_threads));
-          int myStart =  static_cast<int>(gpu_partition + (tid - 1) * worksize);
+          size_t worksize = static_cast<size_t>((storageSize - gpu_partition) / (num_threads));
+          size_t myStart =  static_cast<size_t>(gpu_partition + (tid - 1) * worksize);
           size_t myEnd = static_cast<size_t>(myStart + worksize);
 
           if (tid == num_threads - 1)
@@ -144,7 +144,7 @@ namespace sg {
           //      }
 #if defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t j = (size_t)myStart; j < myEnd; j++) {
+          for (size_t j = myStart; j < myEnd; j++) {
             __m128d res = _mm_set1_pd(0.0f);
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
@@ -286,7 +286,7 @@ namespace sg {
 #endif
 #if defined(__SSE3__) && defined(__AVX__)
 
-          for (size_t j = (size_t)myStart; j < myEnd; j++) {
+          for (size_t j = myStart; j < myEnd; j++) {
             __m256d res = _mm256_set1_pd(0.0f);
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
@@ -432,7 +432,7 @@ namespace sg {
 #endif
 #if !defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t j = (size_t)myStart; j < myEnd; j++) {
+          for (size_t j = myStart; j < myEnd; j++) {
             ptrGlobalResult[j] = 0.0;
 
             for (size_t i = 0; i < source_size; i++) {
@@ -539,14 +539,14 @@ namespace sg {
             gpu_time = omp_get_wtime() - loc_start;
           }
         } else {
-          int worksize = static_cast<int>((result_size - gpu_partition) / (num_threads));
+          size_t worksize = static_cast<size_t>((result_size - gpu_partition) / (num_threads));
 #if defined(__SSE3__) && !defined(__AVX__)
           worksize = (worksize / 8) * 8;
 #endif
 #if defined(__SSE3__) && defined(__AVX__)
           worksize = (worksize / 16) * 16;
 #endif
-          int myStart = static_cast<int>(gpu_partition + (tid - 1) * worksize);
+          size_t myStart = static_cast<size_t>(gpu_partition + (tid - 1) * worksize);
           size_t myEnd = static_cast<size_t>(myStart + worksize);
 
           if (tid == num_threads - 1)
@@ -563,7 +563,7 @@ namespace sg {
           //      }
 #if defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t i = (size_t)myStart; i < myEnd; i += 8) {
+          for (size_t i = myStart; i < myEnd; i += 8) {
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
 
@@ -720,7 +720,7 @@ namespace sg {
 #endif
 #if defined(__SSE3__) && defined(__AVX__)
 
-          for (size_t i = (size_t)myStart; i < myEnd; i += 16) {
+          for (size_t i = myStart; i < myEnd; i += 16) {
             long long imask = 0x7FFFFFFFFFFFFFFF;
             double* fmask = (double*)&imask;
 
@@ -877,7 +877,7 @@ namespace sg {
 #endif
 #if !defined(__SSE3__) && !defined(__AVX__)
 
-          for (size_t i = (size_t)myStart; i < myEnd; i++) {
+          for (size_t i = myStart; i < myEnd; i++) {
             for (size_t j = 0; j < storageSize; j++) {
               double curSupport = ptrAlpha[j];
